riscv64: Add a sample RISCV64 assembly file
Included alongside is the equivalent GAS assembly, with a C test case.
The Go assembly is (obviously) not tested.

For now, the operand ordering is kept consistent with GAS, but we
probably want to switch it up to be consistent with standard Go assembly
instead. See the comment in riscv64.s.

Signed-off-by: Michael Pratt <michael@pratt.im>
Signed-off-by: Michael Scott <mike@foundries.io>