riscv64: cmd/compile/internal: initial RISCV SSA lowering operations
Add initial support for lowering SSA IR to RISCV Progs.

Only the ADD op is added for now, but it is used successfully in the
compiler. Before this change, the compiler complains about failing to
lower NilCheck. With this change it moves on to the next operation.

Note that the opGen.go and rewriteRISCV64.go changes are autogenerated.

Change-Id: I3ae9da011852373966784ba914144c4aacef39d1
Signed-off-by: Michael Scott <mike@foundries.io>