riscv64: cmd/compile/internal/ssa: add dummmy RISCV64 lowering ops
Add lowering ops for all of integer IR operations to lower to ADD
instructions. The compiler gets very upset if many of the more exotic
operations are lowered to ADD. A few of them have basic implementations.

These are enough to get through lowering of the first Go function. The
next thing needed to continue compilation is an implementation of
Proginfo.

opGen.go and rewriteRISCV64.go are generated.

Change-Id: Ic687b49b7aaec89fd28dadd8b16309c007c6dc92
Signed-off-by: Michael Scott <mike@foundries.io>